+++ src/main/java/net/origins/inventive_inventory/commands/config/type/ConfigType.java
@@ -1,8 +1,0 @@
-package net.origins.inventive_inventory.commands.config.type;
-
-public enum ConfigType {
-    SORTING,
-    AUTOMATIC_REFILLING,
-    PROFILES,
-    LOCKED_SLOTS
-}