+++ src/main/java/net/origins/inventive_inventory/config/options/buttons/EnumButtonOption.java
@@ -1,37 +1,26 @@
-package net.origins.inventive_inventory.config.options;
+package net.origins.inventive_inventory.config.options.buttons;
 
+import net.minecraft.client.gui.tooltip.Tooltip;
 import net.minecraft.client.gui.widget.CyclingButtonWidget;
-import net.minecraft.client.option.SimpleOption;
 import net.minecraft.text.Text;
 import net.origins.inventive_inventory.InventiveInventory;
-import net.origins.inventive_inventory.commands.config.type.ConfigType;
 import net.origins.inventive_inventory.config.enums.accessors.Translatable;
+import net.origins.inventive_inventory.config.options.ConfigOption;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Arrays;
 
-public class AdvancedConfigOption<E extends Enum<E>> extends ConfigOption<E> {
+public class EnumButtonOption<E extends Enum<E>> extends ConfigOption<E> {
     private final Class<E> enumClass;
-    private final Text tooltip;
 
-    public AdvancedConfigOption(String key, E defaultValue, ConfigType configType) {
-        this(key, key, defaultValue, configType);
-    }
-
-    public AdvancedConfigOption(String key, String tooltipKey, E defaultValue, ConfigType configType) {
-        super(key, defaultValue, configType);
+    public EnumButtonOption(String tab, String translationKey, E defaultValue) {
+        super(tab, translationKey, defaultValue);
         this.enumClass = defaultValue.getDeclaringClass();
-        this.tooltip = Text.translatable("config.option_tooltip." + InventiveInventory.MOD_ID + "." + tooltipKey);
     }
 
     protected void cycle() {
         E[] values = enumClass.getEnumConstants();
         this.setValue(values[(this.getValue().ordinal() + 1) % values.length]);
-    }
-
-    @Override
-    public E[] getValues() {
-        return enumClass.getEnumConstants();
     }
 
     @Override
@@ -41,7 +30,7 @@ public class AdvancedConfigOption<E extends Enum<E>> extends ConfigOption<E> {
                 this.setValue(config);
                 return;
             } else if (config instanceof Translatable) {
-                if (((Translatable) config).getText().getString().equalsIgnoreCase(value)) {
+                if (((Translatable) config).getButtonText().getString().equalsIgnoreCase(value)) {
                     this.setValue(config);
                     return;
                 }
@@ -50,11 +39,12 @@ public class AdvancedConfigOption<E extends Enum<E>> extends ConfigOption<E> {
     }
 
     @Override
-    public CyclingButtonWidget<?> asButton() {
+    public CyclingButtonWidget<?> asWidget() {
         return CyclingButtonWidget.builder(ConfigOption::getValueAsText)
-                .tooltip(SimpleOption.constantTooltip(this.tooltip))
+                .tooltip(value -> Tooltip.of(Text.translatable("config." + this.tab + ".button.tooltip." + InventiveInventory.MOD_ID + "." + ((Translatable) value).getTranslationKey())))
                 .omitKeyText()
                 .values(Arrays.stream(this.enumClass.getEnumConstants()).toArray()).initially(this.getValue())
                 .build(Text.empty(), (button, value) -> this.cycle());
     }
+
 }