+++ src/main/java/net/origins/inventive_inventory/features/locked_slots/LockedSlotsHandler.java
@@ -20,15 +20,24 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+import java.util.function.BiConsumer;
 import java.util.stream.Collectors;
 
 public class LockedSlotsHandler {
     private static final String LOCKED_SLOTS_FILE = "locked_slots.json";
     public static final Path LOCKED_SLOTS_PATH = ConfigManager.CONFIG_PATH.resolve(LOCKED_SLOTS_FILE);
-    private static LockedSlots lockedSlots = new LockedSlots(List.of());
     private static final List<ItemStack> savedInventory = new ArrayList<>();
     private static final List<ItemStack> savedHandlerInventory = new ArrayList<>();
     public static boolean shouldAdd;
+    public static boolean shouldInit;
+    public static boolean schedulerStarted;
+    private static LockedSlots lockedSlots = new LockedSlots(List.of());
+
+    public static final int HOVER_COLOR = 0x66FF0000;
+    public static final int LOCKED_HOVER_COLOR = 0xFF8B0000;
 
     public static void toggle(int slot) {
         ContextManager.setContext(Contexts.LOCKED_SLOTS);
@@ -55,7 +64,9 @@ public class LockedSlotsHandler {
         }
     }
 
-    public static void initLockedSlots() {
+    public static void init() {
+        reset();
+        shouldInit = false;
         JsonElement jsonFile = FileHandler.get(LOCKED_SLOTS_PATH);
         JsonArray lockedSlotsJson = new JsonArray();
         if (jsonFile.isJsonObject() && jsonFile.getAsJsonObject().has(InventiveInventory.getWorldName())) {
@@ -66,25 +77,27 @@ public class LockedSlotsHandler {
         }
     }
 
-    public static LockedSlots getLockedSlots() {
-        return lockedSlots.adjust();
+    public static void reset() {
+        lockedSlots.clear();
+        savedInventory.clear();
+        savedHandlerInventory.clear();
     }
 
-    public static void clearLockedSlots() {
-        lockedSlots.clear();
+    public static LockedSlots getLockedSlots() {
+        return lockedSlots.adjust();
     }
 
     public static void adjustInventory() {
         List<ItemStack> currentInventory = new ArrayList<>(InventiveInventory.getPlayer().getInventory().main);
         if (savedInventory.isEmpty()) return;
-        boolean tookItem = tookItem(currentInventory);
-        boolean movedItem = movedItem(tookItem);
+        boolean itemTaken = isItemTaken(currentInventory);
+        boolean itemMoved = isItemMoved(itemTaken);
 
         ContextManager.setContext(Contexts.LOCKED_SLOTS);
-        if (ConfigManager.PICKUP_INTO_LOCKED_SLOTS.is(false) && tookItem && !movedItem) {
-            rearrangeThenDrop(currentInventory);
-        } else if (ConfigManager.QUICK_MOVE_INTO_LOCKED_SLOTS.is(false) && movedItem) {
-            rearrangeThenPutBack(currentInventory);
+        if (ConfigManager.PICKUP_INTO_LOCKED_SLOTS.is(false) && itemTaken && !itemMoved) {
+            rearrange(currentInventory, InteractionHandler::dropItem);
+        } else if (ConfigManager.QUICK_MOVE_INTO_LOCKED_SLOTS.is(false) && itemMoved) {
+            rearrange(currentInventory, (slot, times) -> InteractionHandler.quickMove(slot));
         }
         ContextManager.setContext(Contexts.INIT);
     }
@@ -104,30 +117,30 @@ public class LockedSlotsHandler {
         }
     }
 
-    private static boolean tookItem(List<ItemStack> currentInventory) {
-        Map<Item, Integer> currentCountMap = currentInventory.stream()
-                .collect(Collectors.toMap(ItemStack::getItem, ItemStack::getCount, Integer::sum));
-        Map<Item, Integer> savedCountMap = savedInventory.stream()
-                .collect(Collectors.toMap(ItemStack::getItem, ItemStack::getCount, Integer::sum));
+    private static Map<Item, Integer> getItemCountMap(List<ItemStack> inventory) {
+        return inventory.stream().collect(Collectors.toMap(ItemStack::getItem, ItemStack::getCount, Integer::sum));
+    }
+
+    private static boolean isItemTaken(List<ItemStack> currentInventory) {
+        Map<Item, Integer> currentCountMap = getItemCountMap(currentInventory);
+        Map<Item, Integer> savedCountMap = getItemCountMap(savedInventory);
         return currentCountMap.entrySet().stream()
                 .anyMatch(entry -> entry.getValue() > savedCountMap.getOrDefault(entry.getKey(), 0));
     }
 
-    private static boolean movedItem(boolean tookItem) {
+    private static boolean isItemMoved(boolean tookItem) {
         List<ItemStack> currentHandlerInventory = new ArrayList<>();
         for (int i = 0; i < InventiveInventory.getScreenHandler().slots.size(); i++) {
             currentHandlerInventory.add(InteractionHandler.getStackFromSlot(i).copy());
         }
-
-        Map<Item, Integer> currentCountMap = currentHandlerInventory.stream()
-                .collect(Collectors.toMap(ItemStack::getItem, ItemStack::getCount, Integer::sum));
-        Map<Item, Integer> savedCountMap = savedHandlerInventory.stream()
-                .collect(Collectors.toMap(ItemStack::getItem, ItemStack::getCount, Integer::sum));
+        Map<Item, Integer> currentCountMap = getItemCountMap(currentHandlerInventory);
+        Map<Item, Integer> savedCountMap = getItemCountMap(savedHandlerInventory);
         return tookItem && currentCountMap.entrySet().stream()
                 .allMatch(entry -> entry.getValue().equals(savedCountMap.getOrDefault(entry.getKey(), 0)));
     }
 
-    private static void rearrangeThenDrop(List<ItemStack> currentInventory) {
+    private static void rearrange(List<ItemStack> currentInventory, BiConsumer<Integer, Integer> func) {
+        if (savedInventory.isEmpty()) return;
         LockedSlots lockedSlots = LockedSlotsHandler.getLockedSlots();
         int i = 9;
         for (int invSlot : PlayerSlots.get()) {
@@ -155,40 +168,7 @@ public class LockedSlotsHandler {
                 InteractionHandler.leftClickStack(invSlot);
             }
             if (InteractionHandler.getStackFromSlot(invSlot).getCount() > savedStack.getCount()) {
-                InteractionHandler.dropItem(invSlot, InteractionHandler.getStackFromSlot(invSlot).getCount() - savedStack.getCount());
-            }
-        }
-    }
-
-    private static void rearrangeThenPutBack(List<ItemStack> currentInventory) {
-        LockedSlots lockedSlots = LockedSlotsHandler.getLockedSlots();
-        int i = 9;
-        for (int invSlot : PlayerSlots.get()) {
-            ItemStack currentStack = currentInventory.get(i);
-            ItemStack savedStack = savedInventory.get(i);
-            i++;
-            if (!lockedSlots.contains(invSlot) || ItemStack.areEqual(currentStack, savedStack)) continue;
-            List<Integer> suitableSlots = PlayerSlots.get().append(SlotTypes.HOTBAR).exclude(SlotTypes.LOCKED_SLOT).stream()
-                    .filter(slot -> {
-                        ItemStack stack = InteractionHandler.getStackFromSlot(slot);
-                        return stack.isEmpty() || ItemStack.areItemsEqual(stack, currentStack) && stack.getCount() < stack.getMaxCount();
-                    })
-                    .sorted(Comparator.comparing((Integer slot) -> InteractionHandler.getStackFromSlot(slot).getCount(), Comparator.reverseOrder())
-                            .thenComparing(slot -> slot))
-                    .toList();
-            if (!suitableSlots.isEmpty()) {
-                InteractionHandler.leftClickStack(invSlot);
-                for (int slot : suitableSlots) {
-                    ItemStack stack = InteractionHandler.getStackFromSlot(slot);
-                    while (InteractionHandler.getCursorStack().getCount() > savedStack.getCount()) {
-                        if (stack.getCount() < stack.getMaxCount()) InteractionHandler.rightClickStack(slot);
-                        else break;
-                    }
-                }
-                InteractionHandler.leftClickStack(invSlot);
-            }
-            if (InteractionHandler.getStackFromSlot(invSlot).getCount() > savedStack.getCount()) {
-                InteractionHandler.quickMove(invSlot);
+                func.accept(invSlot, InteractionHandler.getStackFromSlot(invSlot).getCount() - savedStack.getCount());
             }
         }
     }
@@ -201,5 +181,23 @@ public class LockedSlotsHandler {
         jsonObject.remove(InventiveInventory.getWorldName());
         jsonObject.add(InventiveInventory.getWorldName(), lockedSlotsJson);
         FileHandler.write(LOCKED_SLOTS_PATH, jsonObject);
+    }
+
+    public static void startScheduler() {
+        ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1);
+        Runnable task = new Runnable() {
+            private int iteration = 0;
+
+            @Override
+            public void run() {
+                if (iteration > 10) {
+                    LockedSlotsHandler.shouldInit = true;
+                    scheduler.shutdown();
+                }
+                iteration++;
+            }
+        };
+        scheduler.scheduleAtFixedRate(task, 0, 50, TimeUnit.MILLISECONDS);
+        schedulerStarted = true;
     }
 }