+++ src/main/java/net/origins/inventive_inventory/util/ScreenCheck.java
@@ -2,9 +2,8 @@ package net.origins.inventive_inventory.util;
 
 import net.minecraft.client.gui.screen.ingame.CreativeInventoryScreen;
 import net.minecraft.client.gui.screen.ingame.InventoryScreen;
-import net.minecraft.screen.CrafterScreenHandler;
-import net.minecraft.screen.PlayerScreenHandler;
-import net.minecraft.screen.ScreenHandler;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.screen.*;
 import net.origins.inventive_inventory.InventiveInventory;
 import net.origins.inventive_inventory.features.profiles.gui.ProfilesScreen;
 
@@ -29,12 +28,13 @@ public class ScreenCheck {
         return InventiveInventory.getScreenHandler() instanceof PlayerScreenHandler;
     }
 
-    public static boolean hasExtraSlot() {
-        ScreenHandler screenHandler = InventiveInventory.getScreenHandler();
-        return screenHandler instanceof PlayerScreenHandler || screenHandler instanceof CrafterScreenHandler;
-    }
-
     public static boolean isProfileScreen() {
         return InventiveInventory.getScreen() instanceof ProfilesScreen;
+    }
+
+    public static boolean isContainer() {
+        ScreenHandler screenHandler = InventiveInventory.getScreenHandler();
+        return screenHandler instanceof GenericContainerScreenHandler || screenHandler instanceof Generic3x3ContainerScreenHandler
+                || screenHandler instanceof HopperScreenHandler || screenHandler.slots.size() - PlayerInventory.MAIN_SIZE > 10;
     }
 }