+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
 
-### Bugs
-- Fixed fps-drops in screens when no locked slot was set
+### General
+- Added compatibility with Trinkets Mods
+- Added compatibility of Locked Slots with many mod screens
+- Added compatibility with Auto Tools
+- Improved performance
+- Multi-Server networks are now supported
+
+### Sorting
+- Enchantments of enchanted items will be taken into account for sorting
+
+## Bug Fixes
+- Fixed sorting in some screens